qiskit_backend/requirements.txt
flask==2.2.2
flask-cors==4.0.0
qiskit==0.44.1
matplotlib==3.7.2
numpy==1.21.5
python-dotenv==1.0.0
